backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.4.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.32.0
basicsr==1.4.2
realesrgan==0.3.0
facexlib==0.3.0

